Commit on 10th April: Leetcode: 20. Valid Parentheses Solution: Use a stack to keep values, and a map to check the storage values to popp when we get a match
Leetcode: 1857. Largest Color Value in a Directed Graph
Solution:
Check for the largest frequrency of color values in a directed path of graph. Use a 2-D matrix to check the color values for the index positions on a specific route and then take the max possible value in the matrix.